requirements.txt
Django>=1.8,<1.9
django-rest-auth>=0.4,<0.5
djangorestframework>=3.1,<3.2
django-cors-headers>=1.1.0
PyYAML
psycopg2>=2.0,<3.0

# development deps
mypy-lang
django-rest-swagger
django-grappelli
pep8
